pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["Faisal Latif <faisal.latif1295@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
mypy = "^1.9.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
dev="todos.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
